[frontend] Code refactoring
